Use autoconf-dickey instead of mainline autoconf
mawk (and other programs maintained upstream by Dickey) is not
compatible with the mainline autoconf, eg. AC_DIVERT_HELP
is only in his version.

Add the build-dependency and override the program used by
dh_autoreconf (while still passing the default flags '-f -i'
which we now have to do explicitly as we override the program).